requirements.txt
Flask
flask-cors
python-dotenv
gensim
nltk